Move indexer implementation to index.processor subpackage (for Jigsaw)
Issue: SPR-16979
