+++ spring-context-indexer/src/main/java/org/springframework/context/index/processor/PackageInfoStereotypesProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2017 the original author or authors.
+ * Copyright 2002-2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,26 +14,32 @@
  * limitations under the License.
  */
 
-package org.springframework.context.index;
+package org.springframework.context.index.processor;
 
+import java.util.HashSet;
 import java.util.Set;
 import javax.lang.model.element.Element;
+import javax.lang.model.element.ElementKind;
 
 /**
- * Provide the list of stereotypes that match an {@link Element}.
- * If an element has one more stereotypes, it is referenced in the index
- * of candidate components and each stereotype can be queried individually.
+ * A {@link StereotypesProvider} implementation that provides the
+ * {@value STEREOTYPE} stereotype for each package-info.
  *
  * @author Stephane Nicoll
  * @since 5.0
  */
-interface StereotypesProvider {
+class PackageInfoStereotypesProvider implements StereotypesProvider {
 
-	/**
-	 * Return the stereotypes that are present on the given {@link Element}.
-	 * @param element the element to handle
-	 * @return the stereotypes or an empty set if none were found
-	 */
-	Set<String> getStereotypes(Element element);
+	public static final String STEREOTYPE = "package-info";
+
+
+	@Override
+	public Set<String> getStereotypes(Element element) {
+		Set<String> stereotypes = new HashSet<>();
+		if (element.getKind() == ElementKind.PACKAGE) {
+			stereotypes.add(STEREOTYPE);
+		}
+		return stereotypes;
+	}
 
 }
