+++ spring-context-indexer/src/main/java/org/springframework/context/index/processor/TypeHelper.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.context.index;
+package org.springframework.context.index.processor;
 
 import java.util.ArrayList;
 import java.util.List;
