+++ spring-context-indexer/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-org.springframework.context.index.CandidateComponentsIndexer
+org.springframework.context.index.processor.CandidateComponentsIndexer
